Got rid of queries package, and moved Mongo-centric files into a package.
