+++ src/main/java/com/spotifydb/ui/controllers/AutocompleteController.java
@@ -1,13 +1,11 @@
 package com.spotifydb.ui.controllers;
 
-import org.bson.Document;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import com.spotifydb.application.DatabaseService;
 
 
-import java.util.List;
 
 @Controller
 @RequestMapping("/autocomplete")
